+++ COFantasy.js
@@ -1297,6 +1297,43 @@ var COFantasy = COFantasy || function() {
     }
   }
 
+  //Retourne le perso correspondant à un token id suivi du nom de token
+  //Permet d'avoir une information robuste en cas d'interruption du script
+  function persoOfIdName(idn, pageId) {
+    var pos = idn.indexOf(' ');
+    if (pos < 1 || pos >= idn.length) {
+      error("IdName mal formé", idn);
+      return;
+    }
+    var name = idn.substring(pos + 1);
+    var perso = persoOfId(idn.substring(0, pos), name, pageId);
+    perso.tokName = perso.token.get('name');
+    if (perso.tokName == name) return perso;
+    log("En cherchant le token " + idn + ", on trouve " + perso.tokName);
+    log(perso);
+    return perso;
+  }
+
+  //options peut avoir les champs:
+  // - strict1 = true si on considère que tok1 doit avoir une taille nulle
+  // - strict2
+  // - allonge
+  function distanceCombat(tok1, tok2, pageId, options) {
+    if (pageId === undefined) {
+      pageId = tok1.get('pageid');
+    }
+    var scale = computeScale(pageId);
+    var pt1 = tokenCenter(tok1);
+    var pt2 = tokenCenter(tok2);
+    var distance_pix = VecMath.length(VecMath.vec(pt1, pt2));
+    options = options || {};
+    if (!options.strict1) distance_pix -= tokenSize(tok1, PIX_PER_UNIT / 2);
+    if (!options.strict2) distance_pix -= tokenSize(tok2, PIX_PER_UNIT / 2);
+    if (options.allonge) distance_pix -= (options.allonge * PIX_PER_UNIT) / scale;
+    if ((!options.strict1 || !options.strict2) && distance_pix < PIX_PER_UNIT * 1.3) return 0; //cases voisines
+    return ((distance_pix / PIX_PER_UNIT) * scale);
+  }
+
   function setState(personnage, etat, value, evt) {
     var token = personnage.token;
     var charId = personnage.charId;
@@ -1304,7 +1341,7 @@ var COFantasy = COFantasy || function() {
       affectToken(token, 'statusmarkers', token.get('statusmarkers'), evt);
     if (value && etatRendInactif(etat) && isActive(personnage)) {
       if (etat != 'surpris' || !surveillance(personnage))
-        removeFromTurnTracker(token.id, evt);
+        removeFromTurnTracker(personnage, evt);
     }
     token.set(cof_states[etat], value);
     if (!value) { //On enlève le save si il y en a un
@@ -1362,7 +1399,7 @@ var COFantasy = COFantasy || function() {
         a.remove();
       });
       if (charAttributeAsBool(personnage, 'armeeConjuree')) {
-        removeFromTurnTracker(personnage.token.id, evt);
+        removeFromTurnTracker(personnage, evt);
         personnage.token.remove();
         sendChar(personnage.charId, 'disparaît');
         var armeeChar = getObj('character', personnage.charId);
@@ -1803,21 +1840,6 @@ var COFantasy = COFantasy || function() {
       token: token,
       charId: charId
     };
-  }
-
-  function persoOfIdName(idn, pageId) {
-    var pos = idn.indexOf(' ');
-    if (pos < 1 || pos >= idn.length) {
-      error("IdName mal formé", idn);
-      return;
-    }
-    var name = idn.substring(pos + 1);
-    var perso = persoOfId(idn.substring(0, pos), name, pageId);
-    perso.tokName = perso.token.get('name');
-    if (perso.tokName == name) return perso;
-    log("En cherchant le token " + idn + ", on trouve " + perso.tokName);
-    log(perso);
-    return perso;
   }
 
   function boutonSimple(action, style, texte) {
@@ -2559,25 +2581,6 @@ var COFantasy = COFantasy || function() {
         break;
     }
     return scale;
-  }
-  //options peut avoir les champs:
-  // - strict1 = true si on considère que tok1 doit avoir une taille nulle
-  // - strict2
-  // - allonge
-  function distanceCombat(tok1, tok2, pageId, options) {
-    if (pageId === undefined) {
-      pageId = tok1.get('pageid');
-    }
-    var scale = computeScale(pageId);
-    var pt1 = tokenCenter(tok1);
-    var pt2 = tokenCenter(tok2);
-    var distance_pix = VecMath.length(VecMath.vec(pt1, pt2));
-    options = options || {};
-    if (!options.strict1) distance_pix -= tokenSize(tok1, PIX_PER_UNIT / 2);
-    if (!options.strict2) distance_pix -= tokenSize(tok2, PIX_PER_UNIT / 2);
-    if (options.allonge) distance_pix -= (options.allonge * PIX_PER_UNIT) / scale;
-    if ((!options.strict1 || !options.strict2) && distance_pix < PIX_PER_UNIT * 1.4) return 0; //cases voisines
-    return ((distance_pix / PIX_PER_UNIT) * scale);
   }
 
   function getPageId(playerId) {
@@ -5668,13 +5671,36 @@ var COFantasy = COFantasy || function() {
         }
       }
     }
-    var attaqueEnMeute = charAttributeAsInt(attaquant, 'attaqueEnMeute', 0);
-    if (attaqueEnMeute > 0) {
-      if (attributeAsBool(target, 'attaqueParMeute')) {
-        attBonus += attaqueEnMeute;
-        explications.push("Attaque en meute => +" + attaqueEnMeute + " pour toucher");
+    if (options.attaqueEnMeute) {
+      var attaqueParMeute = tokenAttribute(target, 'attaqueParMeute');
+      if (attaqueParMeute.length > 0) {
+        attaqueParMeute = attaqueParMeute[0];
+        var attaqueParMeuteCur = attaqueParMeute.get('current');
+        var contientAttaquant;
+        var autreAttaquant;
+        attaqueParMeuteCur.split(' ').forEach(function(mi) {
+          if (mi == attaquant.token.id) {
+            contientAttaquant = true;
+            return;
+          }
+          autreAttaquant = true;
+        });
+        if (autreAttaquant) {
+          attBonus += options.attaqueEnMeute;
+          explications.push("Attaque en meute => +" + options.attaqueEnMeute + " pour toucher");
+        }
+        if (!contientAttaquant) {
+          evt.attributes = evt.attributes || [];
+          evt.attributes.push({
+            attribute: attaqueParMeute,
+            current: attaqueParMeuteCur
+          });
+          if (attaqueParMeuteCur === '') attaqueParMeuteCur = attaquant.token.id;
+          else attaqueParMeuteCur += ' ' + attaquant.token.id;
+          attaqueParMeute.set('current', attaqueParMeuteCur);
+        }
       } else {
-        setTokenAttr(target, 'attaqueParMeute', true, evt);
+        setTokenAttr(target, 'attaqueParMeute', attaquant.token.id, evt);
       }
     }
     return attBonus;
@@ -6269,7 +6295,6 @@ var COFantasy = COFantasy || function() {
         });
       }
     }
-
     //Les conditions qui peuvent empêcher l'attaque
     if (options.conditionAttaquant !== undefined) {
       if (!testCondition(options.conditionAttaquant, attaquant, cibles)) {
@@ -6281,10 +6306,11 @@ var COFantasy = COFantasy || function() {
       sendChar(attackingCharId, "ne peut pas utiliser cette capacité quand il est affaibli.");
       return;
     }
-    var attrRipostesDuTour = tokenAttribute(attaquant, 'ripostesDuTour');
+    //dernieresCiblesAttaquees contient en current les cibles attaquées, et en max les cibles sur lesquelles on a fait des ripostes
+    var attrCiblesAttaquees = tokenAttribute(attaquant, 'dernieresCiblesAttaquees');
     var ripostesDuTour = new Set();
-    if (attrRipostesDuTour.length > 0) {
-      ripostesDuTour = new Set(attrRipostesDuTour[0].get('current').split(' '));
+    if (attrCiblesAttaquees.length > 0) {
+      ripostesDuTour = new Set(attrCiblesAttaquees[0].get('max').split(' '));
     }
     cibles = cibles.filter(function(target) {
       if (attributeAsBool(target, 'ombreMortelle')) {
@@ -6388,36 +6414,43 @@ var COFantasy = COFantasy || function() {
       options.messages = options.messages || [];
       degainerArme(attaquant, attackLabel, evt, options);
     }
-    if (charAttributeAsBool(attaquant, 'riposte')) {
-      //On stoque les cibles attaquées, pour ne pas les re-proposer en riposte
+    var riposte = charAttributeAsBool(attaquant, 'riposte');
+    options.attaqueEnMeute = charAttributeAsInt(attaquant, 'attaqueEnMeute', 0);
+    var attrLienEpique = charAttribute(attaquant.charId, 'lienEpique');
+    if (attrLienEpique.length > 0) {
+      options.lienEpique = attrLienEpique[0].get('current');
+    }
+    if (riposte || options.attaqueEnMeute || options.lienEpique) {
+      //Dans ce cas, il faut stoquer les cibles attaquées
+      //(dans le cas de riposte, pour ne pas les re-proposer en riposte
       var listeCibles =
         cibles.map(function(target) {
           return target.token.id;
         }).join(' ');
-      if (attrRipostesDuTour.length === 0) {
+      if (attrCiblesAttaquees.length === 0) {
         if (options.riposte) {
-          setTokenAttr(attaquant, 'ripostesDuTour', listeCibles, evt);
+          setTokenAttr(attaquant, 'dernieresCiblesAttaquees', '', evt, undefined, listeCibles);
         } else {
-          setTokenAttr(attaquant, 'ripostesDuTour', '', evt, undefined, listeCibles);
+          setTokenAttr(attaquant, 'dernieresCiblesAttaquees', listeCibles, evt);
         }
       } else { //L'attribut existe déjà
-        attrRipostesDuTour = attrRipostesDuTour[0];
+        attrCiblesAttaquees = attrCiblesAttaquees[0];
         evt.attributes = evt.attributes || [];
-        ripostesDuTour = attrRipostesDuTour.get('current');
-        var attaquesDuTour = attrRipostesDuTour.get('max');
+        var attaquesDuTour = attrCiblesAttaquees.get('current');
+        ripostesDuTour = attrCiblesAttaquees.get('max');
         evt.attributes.push({
-          attribute: attrRipostesDuTour,
-          current: ripostesDuTour,
-          max: attaquesDuTour
+          attribute: attrCiblesAttaquees,
+          current: attaquesDuTour,
+          max: ripostesDuTour,
         });
         if (options.riposte) {
           if (ripostesDuTour === '') ripostesDuTour = listeCibles;
           else ripostesDuTour += ' ' + listeCibles;
-          attrRipostesDuTour.set('current', ripostesDuTour);
+          attrCiblesAttaquees.set('max', ripostesDuTour);
         } else {
           if (attaquesDuTour === '') attaquesDuTour = listeCibles;
           else attaquesDuTour += ' ' + listeCibles;
-          attrRipostesDuTour.set('max', attaquesDuTour);
+          attrCiblesAttaquees.set('current', attaquesDuTour);
         }
       }
     }
@@ -7649,7 +7682,6 @@ var COFantasy = COFantasy || function() {
     var attackingToken = attaquant.token;
     var attackerTokName = attaquant.tokName;
     options.attaquant = attaquant;
-
     //Les dégâts
     //Dégâts insrits sur la ligne de l'arme
     var mainDmgType = options.type || 'normal';
@@ -7861,6 +7893,27 @@ var COFantasy = COFantasy || function() {
         }
       }
     };
+    //Le lien épique (+1d6 DM si les 2 attaquent la même cible
+    var attaqueParLienEpique = new Set();
+    if (options.lienEpique) {
+      //On cherche les autres personnages avec le même lien épique
+      var allChars = findObjs({
+        type: 'character'
+      });
+      allChars.forEach(function(ch) {
+        if (ch.id == attackingCharId) return;
+        var attrLienEpique = charAttribute(ch.id, 'lienEpique');
+        if (attrLienEpique.length === 0) return;
+        if (attrLienEpique[0].get('current') != options.lienEpique) return;
+        var attrCibles = charAttribute(ch.id, 'dernieresCiblesAttaquees');
+        if (attrCibles.length === 0) return;
+        var ciblesAttaquees = attrCibles[0].get('current');
+        if (ciblesAttaquees === '') return;
+        ciblesAttaquees.split(' ').forEach(function(ci) {
+          attaqueParLienEpique.add(ci);
+        });
+      });
+    }
     cibles.forEach(function(target) {
       evalITE(attaquant, target, d20roll, options, evt, explications, options);
       target.ignoreTouteRD = target.ignoreTouteRD || options.ignoreTouteRD;
@@ -7962,6 +8015,13 @@ var COFantasy = COFantasy || function() {
           type: mainDmgType,
           value: '1d6'
         });
+      }
+      if (options.lienEpique && attaqueParLienEpique.has(target.token.id)) {
+        target.additionalDmg.push({
+          type: mainDmgType,
+          value: '1' + options.d6
+        });
+        target.messages.push("Lien épique => + 1" + options.d6 + " DM");
       }
       var attrFeinte = tokenAttribute(target, 'feinte_' + attaquant.tokName);
       if (attrFeinte.length > 0 && attrFeinte[0].get('current') &&
@@ -8779,13 +8839,13 @@ var COFantasy = COFantasy || function() {
         if (charAttributeAsBool(target, 'seulContreTous')) {
           displayAttaqueOpportunite(target.token.id, [attaquant], "de riposte", '#ActionsRiposte#');
         } else if (charAttributeAsBool(target, 'riposte')) {
-          var attrRipostesDuTour = tokenAttribute(target, 'ripostesDuTour');
-          if (attrRipostesDuTour.length > 0) {
+          var attrCiblesDeLaCible = tokenAttribute(target, 'dernieresCiblesAttaquees');
+          if (attrCiblesDeLaCible.length > 0) {
             var ripostesDuTour =
-              attrRipostesDuTour[0].get('current').split(' ');
+              attrCiblesDeLaCible[0].get('max').split(' ');
             ripostesDuTour = new Set(ripostesDuTour);
             if (ripostesDuTour.has(attaquant.token.id)) return;
-            ripostesDuTour = attrRipostesDuTour[0].get('max').split(' ');
+            ripostesDuTour = attrCiblesDeLaCible[0].get('current').split(' ');
             ripostesDuTour = new Set(ripostesDuTour);
             if (ripostesDuTour.has(attaquant.token.id)) return;
           }
@@ -10299,6 +10359,7 @@ var COFantasy = COFantasy || function() {
     attrs = removeAllAttributes('attaqueMalgreMenace', evt, attrs);
     attrs = removeAllAttributes('limiteApplicationManoeuvre', evt, attrs);
     attrs = removeAllAttributes('attaqueParMeute', evt, attrs);
+    attrs = removeAllAttributes('dernieresCiblesAttaquees', evt, attrs);
     // Autres attributs
     // Remettre le pacifisme au max
     resetAttr(attrs, 'pacifisme', evt, "retrouve son pacifisme");
@@ -10933,7 +10994,7 @@ var COFantasy = COFantasy || function() {
             error("Il manque le nom de l'imageaprès --image", cmd);
             return;
           }
-          options.image = cmd[1];
+          options.image = cmd[1].replace('&#58;', ':');
           return;
         case 'son':
           if (cmd.length < 2) {
@@ -12991,6 +13052,50 @@ var COFantasy = COFantasy || function() {
     });
   }
 
+  function removeDernieresCiblesAttaquees(perso, evt) {
+    var attrDernieresCibles = tokenAttribute(perso, 'dernieresCiblesAttaquees');
+    if (attrDernieresCibles.length > 0) {
+      attrDernieresCibles = attrDernieresCibles[0];
+      if (charAttributeAsBool(perso, 'attaqueEnMeute')) {
+        var dernieresCibles = attrDernieresCibles.get('current');
+        var cibles = new Set(dernieresCibles.split(' '));
+        cibles.forEach(function(ci) {
+          var cible = persoOfId(ci);
+          if (cible === undefined) return;
+          var attrCibleMeute = tokenAttribute(cible, 'attaqueParMeute');
+          if (attrCibleMeute.length > 0) {
+            attrCibleMeute = attrCibleMeute[0];
+            var cibleMeute = attrCibleMeute.get('current');
+            var ensembleCibleMeute = new Set(cibleMeute.split(' '));
+            if (ensembleCibleMeute.has(perso.token.id)) {
+              ensembleCibleMeute.delete(perso.token.id);
+              if (ensembleCibleMeute.size > 0) {
+                evt.attributes = evt.attributes || [];
+                evt.attributes.push({
+                  attribute: attrCibleMeute,
+                  current: cibleMeute,
+                });
+                cibleMeute = '';
+                ensembleCibleMeute.forEach(function(ai) {
+                  if (cibleMeute === '') cibleMeute = ai;
+                  else cibleMeute += ' ' + ai;
+                });
+                attrCibleMeute.set('current', cibleMeute);
+              } else {
+                evt.deletedAttributes = evt.deletedAttributes || [];
+                evt.deletedAttributes.push(attrCibleMeute);
+                attrCibleMeute.remove();
+              }
+            }
+          }
+        });
+      }
+      evt.deletedAttributes = evt.deletedAttributes || [];
+      evt.deletedAttributes.push(attrDernieresCibles);
+      attrDernieresCibles.remove();
+    }
+  }
+
   function setActiveToken(tokenId, evt) {
     var pageId = Campaign().get('initiativepage');
     if (stateCOF.activeTokenId) {
@@ -13032,6 +13137,8 @@ var COFantasy = COFantasy || function() {
     if (tokenId) {
       var perso = persoOfId(tokenId, tokenId);
       if (perso) {
+        //On remet à 0 la liste des cibles attaquées par le personnage
+        removeDernieresCiblesAttaquees(perso, evt);
         var token = perso.token;
         var charId = perso.charId;
         // personnage lié au Token
@@ -13573,7 +13680,9 @@ var COFantasy = COFantasy || function() {
     });
   }
 
-  function removeFromTurnTracker(tokenId, evt) {
+  function removeFromTurnTracker(perso, evt) {
+    removeDernieresCiblesAttaquees(perso, evt);
+    var tokenId = perso.token.id;
     var turnOrder = Campaign().get('turnorder');
     if (turnOrder === "" || !stateCOF.combat) {
       return;
@@ -14310,8 +14419,8 @@ var COFantasy = COFantasy || function() {
         }
         if (options && options.messages) message += "rengaine " + ancienneArme.name + " et ";
         else sendChar(perso.charId, "rengaine " + ancienneArme.name);
-        if (charAttributeAsBool(perso, 'eclaire_'+labelArmeActuelle)) {
-          eteindreUneLumiere(perso, pageId, undefined, 'eclaire_'+labelArmeActuelle, evt);
+        if (charAttributeAsBool(perso, 'eclaire_' + labelArmeActuelle)) {
+          eteindreUneLumiere(perso, pageId, undefined, 'eclaire_' + labelArmeActuelle, evt);
         }
       }
     } else armeActuelle = undefined;
@@ -14413,18 +14522,18 @@ var COFantasy = COFantasy || function() {
       message += "dégaine " + nouvelleArme.name;
       options.messages.push(message);
     } else sendChar(perso.charId, "dégaine " + nouvelleArme.name);
-    var eclaire = 'eclaire_'+labelArme;
+    var eclaire = 'eclaire_' + labelArme;
     var attrEclaire = charAttribute(perso.charId, eclaire);
-        if (attrEclaire.length > 0) {
-          var radius = parseInt(attrEclaire[0].get('current'));
-          if (isNaN(radius)) {
-            error("Attribut "+eclaire+" mal formé", attrEclaire);
-          } else {
-          var dimRadius = parseInt(attrEclaire[0].get('max'));
-            if (isNaN(dimRadius)) dimRadius = '';
-          ajouteUneLumiere(perso, eclaire, radius, dimRadius, evt);
-          }
-        }
+    if (attrEclaire.length > 0) {
+      var radius = parseInt(attrEclaire[0].get('current'));
+      if (isNaN(radius)) {
+        error("Attribut " + eclaire + " mal formé", attrEclaire);
+      } else {
+        var dimRadius = parseInt(attrEclaire[0].get('max'));
+        if (isNaN(dimRadius)) dimRadius = '';
+        ajouteUneLumiere(perso, eclaire, radius, dimRadius, evt);
+      }
+    }
     if (charAttributeAsInt(perso, "initEnMain" + labelArme, 0) > 0)
       updateNextInit(perso.token);
   }
@@ -16711,45 +16820,18 @@ var COFantasy = COFantasy || function() {
 
 
   function murDeForce(msg) {
-    var cmd = msg.content.split(' ');
+    var options = parseOptions(msg);
+    if (options === undefined) return;
+    var cmd = options.cmd;
+    if (cmd === undefined) return;
     var sphere = true;
-    if (cmd.length > 1 && cmd[1] == 'mur') sphere = false;
-    var options = {
-      image: stateCOF.options.images.val.image_mur_de_force.val,
-      mana: 0
-    };
-    var args = msg.content.split(' --');
-    args.shift();
-    args.forEach(function(opt) {
-      var optCmd = opt.split(' ');
-      switch (optCmd[0]) {
-        case 'mana':
-          if (optCmd.length < 2) {
-            error("Il manque le coût en mana", cmd);
-            return;
-          }
-          options.mana = parseInt(optCmd[1]);
-          if (isNaN(options.mana) || options.mana < 0) {
-            error("Coût en mana incorrect", optCmd);
-            options.mana = 0;
-          }
-          return;
-        case 'puissant':
-          options.puissant = true;
-          return;
-        case 'image':
-          if (optCmd.length < 2) {
-            error("Il manque l'adresse de l'image", cmd);
-            return;
-          }
-          options.image = optCmd[1].replace('&#58;', ':');
-          return;
-        case 'no-image':
-          options.image = undefined;
-          return;
-      }
-    });
-    getSelected(msg, function(selected) {
+    var imageSphere = stateCOF.options.images.val.image_mur_de_force.val;
+    if (cmd.length > 1) {
+      if (cmd[1] == 'mur') sphere = false;
+      else if (cmd[1] == 'noImage') imageSphere = undefined;
+      else imageSphere = cmd[1].replace('&#58;', ':');
+    }
+    getSelected(msg, function(selected, playerId) {
       if (selected.length === 0) {
         sendPlayer(msg, "Aucun personnage sélectionné pour lancer le mur de force");
         return;
@@ -16759,19 +16841,40 @@ var COFantasy = COFantasy || function() {
       };
       initiative(selected, evt);
       iterSelected(selected, function(lanceur) {
+        if (options.tempeteDeMana) {
+          if (options.tempeteDeMana.cout === 0) {
+            //On demande de préciser les options
+            var optMana = {
+              mana: options.mana,
+              dm: false,
+              soins: false,
+              duree: true,
+              portee:true,
+              rang: options.rang,
+            };
+            setTempeteDeMana(playerId, lanceur, msg.content, optMana);
+            return;
+          } else {
+            if (options.rang && options.tempeteDeMana.cout > options.rang) {
+              sendChar(lanceur.charId, "Attention, le coût de la tempête de mana (" + options.tempeteDeMana.cout + ") est supérieur au rang du sort");
+            }
+          }
+        }
         var charId = lanceur.charId;
         var token = lanceur.token;
         var pageId = token.get('pageid');
-        if (!depenseMana(lanceur, options.mana, "lancer un mur de force", evt)) {
-          return;
-        }
-        sendChar(charId, "lance un sort de mur de force");
-        if (options.image && sphere) {
+        if (limiteRessources(lanceur, options, 'murDeForce', 'lancer un mur de force', evt)) return;
+        if (options.son) playSound(options.son);
+        whisperChar(charId, "lance un sort de mur de force");
+        if (sphere) {
           var scale = computeScale(pageId);
           var diametre = PIX_PER_UNIT * (6 / scale);
+          if (options.tempeteDeManaPortee) diametre += diametre;
+          if (options.tempeteDeManaIntence)
+            diametre *= (1 + options.tempeteDeManaIntense);
           var imageFields = {
             _pageid: pageId,
-            imgsrc: options.image,
+            imgsrc: imageSphere,
             represents: '',
             left: token.get('left'),
             top: token.get('top'),
@@ -16787,6 +16890,7 @@ var COFantasy = COFantasy || function() {
             toFront(newImage);
             setTokenAttr(lanceur, 'murDeForceId', newImage.id, evt);
             var duree = 5 + modCarac(lanceur, 'CHARISME');
+            if (options.tempeteDeManaDuree) duree += duree;
             setTokenAttr(lanceur, 'murDeForce', duree, evt, undefined, getInit());
           } else {
             error("Impossible de créer l'image " + options.image, imageFields);
@@ -23430,7 +23534,11 @@ var COFantasy = COFantasy || function() {
         if (effet == 'arbreAnime') {
           iterTokensOfAttribute(charId, options.pageId, effet, attrName,
             function(token) {
-              removeFromTurnTracker(token.id, evt);
+              var perso = {
+                token: token,
+                charId: charId
+              };
+              removeFromTurnTracker(perso, evt);
               setToken(token, 'bar1_link', '', evt);
               setToken(token, 'bar1_value', '', evt);
               setToken(token, 'bar1_max', '', evt);
@@ -24001,11 +24109,8 @@ var COFantasy = COFantasy || function() {
       attrs = removeAllAttributes('exemplaire', evt, attrs);
       attrs = removeAllAttributes('peutEtreDeplace', evt, attrs);
       attrs = removeAllAttributes('attaqueMalgreMenace', evt, attrs);
-      attrs = removeAllAttributes('ripostesDuTour', evt, attrs);
       resetAttr(attrs, 'attaqueEnTraitre', evt);
       resetAttr(attrs, 'esquiveAcrobatique', evt);
-      // Enlever attaque en meute gobelin
-      attrs = removeAllAttributes('attaqueParMeute', evt, attrs);
       // Pour défaut dans la cuirasse, on diminue si la valeur est 2, et on supprime si c'est 1
       var defautsDansLaCuirasse = allAttributesNamed(attrs, 'defautDansLaCuirasse');
       defautsDansLaCuirasse.forEach(function(attr) {
@@ -24671,7 +24776,7 @@ on("destroy:handout", function(prev) {
 });
 
 on('ready', function() {
-  var script_version = "2.06";
+  var scriptVersion = "2.07";
   on('add:token', COFantasy.addToken);
   on("change:graphic:statusmarkers", COFantasy.changeMarker);
   on("change:campaign:playerpageid", COFantasy.initAllMarkers);
@@ -24680,7 +24785,7 @@ on('ready', function() {
     tour: 0,
     init: 1000,
     eventId: 0,
-    version: script_version,
+    version: scriptVersion,
   };
   COFantasy.setStateCOF();
   if (state.COFantasy.version === undefined) {
@@ -24792,7 +24897,7 @@ on('ready', function() {
         characterid: c.id,
         name: 'scriptVersion',
         current: true,
-        max: script_version
+        max: scriptVersion
       });
       state.COFantasy.scriptSheets = true;
     });
@@ -24899,7 +25004,7 @@ on('ready', function() {
     });
     log("Mise à jour des ability & macros !cof-lancer-sort effectuée");
   }
-  state.COFantasy.version = script_version;
+  state.COFantasy.version = scriptVersion;
   if (state.COFantasy.options.affichage.val.fiche.val) {
     if (!state.COFantasy.scriptSheets) {
       findObjs({
@@ -24921,7 +25026,7 @@ on('ready', function() {
     COFantasy.changeHandout(hand);
   });
   COF_loaded = true;
-  log("COFantasy " + script_version + " loaded");
+  log("COFantasy " + scriptVersion + " loaded");
 });
 
 on("chat:message", function(msg) {