+++ ChangeLog.md
@@ -1,6 +1,14 @@
 # COFantasy: Historique des changements (depuis la version 1.0)
-## 2.06
+## 2.07
 ### Autres améliorations
+* Utilisation des options standard pour le mur de force
+
+## 2.06
+### Capacités
+* Ajout de la capacité lien épique.
+
+### Autres améliorations
+* Amélioration de la gestion de l'attaque en meute des gobelins.
 * Possibilité qu'une arme fasse de la lumière quand elle est dégainée.
 * Extinction de la lumière quand l'effet d'arme enflammée prend fin
 * Ajout d'une option pour modifier la portée d'une attaque.